20_Outils_Clients/openclaw-immo/backend/requirements.txt
# OpenClaw Immo — Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# HTTP client (scraper)
httpx==0.28.1

# Google Sheets
gspread==6.1.4
google-auth==2.37.0

# WhatsApp : via OpenClaw Gateway (gratuit, pas de dep externe)

# Scheduler
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
